skarchitect/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "skarchitect"
version = "0.1.0"
description = "Sovereign civic participation SDK — proposals, voting, delegation for human-AI republics"
readme = "README.md"
license = "GPL-3.0-or-later"
requires-python = ">=3.11"
authors = [{ name = "smilinTux", email = "admin@skworld.io" }]
classifiers = [
    "Development Status :: 3 - Alpha",
    "License :: OSI Approved :: GNU General Public License v3 or later (GPLv3+)",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries",
]
dependencies = [
    "pydantic>=2.0",
    "pynacl>=1.5.0",
    "click>=8.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0", "pytest-asyncio>=0.23"]
capauth = ["gnupg>=2.3"]

[project.scripts]
skarchitect = "skarchitect.cli:cli"

[tool.hatch.build.targets.wheel]
packages = ["src/skarchitect"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
